filming time

+++ Final Project/Activity/Activity.java
@@ -1,68 +1,53 @@
-// Correct Package Declaration (top of file)
 package Activity;
-
-
 import java.util.Scanner;
-
 
 public abstract class Activity {
    private String name;
-   private int duration; // Duration in minutes
+   private float duration;
 
 
-   public Activity(String name, int duration) {
+   public Activity(String name, float duration) {
        this.name = name;
        this.duration = duration;
    }
 
-
-   // Getter methods
    public String getName() {
        return name;
    }
 
-
-   public int getDuration() {
+   public float getDuration() {
        return duration;
    }
 
-
-   // Abstract method that will be implemented by subclasses
    public abstract void performActivity(Scanner scanner);
 
-
-   // Method to display information about the activity
    public void displayInfo() {
        System.out.println("Activity: " + name);
    }
 
-
-   // Method to wait for Enter key before starting the activity
    public void waitForEnter(Scanner scanner) {
        System.out.println("Press Enter to start the activity...");
-       scanner.nextLine(); // Wait for the user to press Enter
+       scanner.nextLine(); 
    }
 
-
-   // Method to get a valid duration input from the user (moved from subclass)
-   protected int getValidDuration(Scanner scanner) {
-       int duration = 0;
+   protected float getValidDuration(Scanner scanner) {
+       float duration = 0;
        boolean valid = false;
       
        while (!valid) {
-           try {
-               duration = Integer.parseInt(scanner.nextLine()); // Get user input
-               if (duration > 0) {
-                   valid = true; // If the duration is positive, accept it
-               } else {
-                   System.out.print("Please enter a positive number for the duration: ");
-               }
-           } catch (NumberFormatException e) {
-               System.out.print("Invalid input. Please enter a valid number for the duration: ");
-           }
-       }
-       return duration;
-   }
+        try {
+            duration = Float.parseFloat(scanner.nextLine()); 
+            if (duration > 0) {
+                valid = true; 
+            } else {
+                System.out.print("Please enter a positive number for the duration: ");
+            }
+        } catch (NumberFormatException e) {
+            System.out.print("Invalid input. Please enter a valid number for the duration: ");
+        }
+    }
+    return duration;
+    }
 }
